requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
base58==2.1.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.3
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==5.0.0
eth-utils==2.3.1
fastapi==0.112.1
h11==0.14.0
idna==3.7
more-itertools==10.4.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
scalecodec==1.2.11
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
substrate-interface==1.7.10
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
websocket-client==1.8.0
xxhash==3.5.0
